requirements.txt
Flask>=2.0
pandas>=1.4
numpy>=1.21
requests>=2.28
gdown>=4.4
python-dotenv>=0.20
pycaret>=2.3
openai>=0.27
